SWIG: Update autoconf build process for swig wrappers
This changes the way how swig generated scripting language wrappers are
build through autoconf/automake. In essence, the changes should provide
a more robust way to check whether and how wrappers can be build, and
where the language interpreters want the corresponding files to be
installed to. Note, that we are now building a so-called convenience
library in lib/ that will be used to link against the shared libraries
created with swig. This doesn't do any harm for the C++ library and will
not get installed anyway.
Also, this patch splits the build for Python2 and 3 interface wrappers
into two separate entities. By doing so, the package can build both
interfaces concurrently and apply all settings necessary. In the future,
when Python2 support is not essential anymore, it can be more easily
dropped by simply removing the respective path.
Lastly, this adds version information to the scripting language
interfaces, i.e. RNA::VERSION, and RNA.__version__ attributes for Perl 5
and Python, respectively.